MAI: use attribute `Function.bdd` instead of function arg `bdd`
in preparation for removing redundant `bdd` arguments from functions
in `autoref` and `cudd`, because this information is available as the
attribute `Function.bdd`.

Moreover, reducing the number of arguments that the user has to give,
and automating these references reduces the likelihood of the error
of passing the wrong manager in user code.