MAI: do not memoize attributes `cudd.BDD.false` and `true`
because this creates cyclic references between `BDD` and `Function`.
We want Python to garbage collect the `BDD` last.
In order for this to happen, the `BDD` should still be referenced when
all nodes have been dereferenced. This cannot be the case if the `BDD`
itself references nodes (via the attributes `BDD._false` and `_true`).

One possible approach is to let `BDD._false` and `_true` be weak
references. These references will survive for as long as external
references are present. If external references to true and false
come and go without overlaps, then the overall result is the same.

So, it is simpler to remove the memoization. After all, the performance
benefit is likely negligible. Premature optimization.