API: change `bdd.BDD.sat_len` to count only levels in support
Optionally, the total number of variables can be given.
This signature is the same with CUDD.
If no variables are given, then those in `support(u)`
are accounted for during counting.

If `n` is larger than `len(support(u))`, then the result
will be larger, accounting for enumeration over variables
that `u` does not depend on.

The recursion remained unchanged. The only change is that
levels of variables in `support(u)` are indexed independently
from "global" levels. The map that results from this local
indexing is used during recursion to translate global to
local levels, and count with the latter indices.